fixed map display bug
