+++ components/wardmap.tsx
@@ -222,13 +222,8 @@ export default function WardMap({
   }, [selectedWard]);
 
   return (
-    <>
     <section className="relative w-full h-[400px] rounded-xl overflow-hidden shadow-lg">
       <div ref={containerRef} style={{ width: "100%", height: "100%" }} />
     </section>
-    <section className="relative w-full h-[500px] rounded-xl overflow-hidden shadow-lg">
-      <section ref={containerRef} style={{ width: "100%", height: "100%" }} />
-    </section>
-    </>
   );
 }
